pyproject.toml
[project]
name = "upwork-project27nov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "cryptography>=46.0.3",
    "fastapi>=0.122.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pymysql>=1.1.2",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.44",
    "uv>=0.9.13",
]
